--- src/main/Application/ForButtonReceipts.fxml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Controller.ControllerForReceipts">
-   <children>
-      <Label fx:id="Receipts" layoutX="252.0" layoutY="102.0" />
-   </children>
-</AnchorPane>
