+++ src/main/Application/ForProducts.fxml
@@ -5,10 +5,10 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Controller.ControllerForProducts">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Application.ControllerForProducts">
    <children>
       <Label fx:id="AllProducts" layoutX="202.0" layoutY="111.0" text="products:" textAlignment="CENTER" />
-      <Button fx:id="cancelButton" cancelButton="true" layoutX="28.0" layoutY="21.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="30.0" text="&lt;">
+      <Button fx:id="cancelButton" onAction = "#goToHome" cancelButton="true" layoutX="28.0" layoutY="21.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="30.0" text="&lt;">
          <font>
             <Font name="Arial Black" size="15.0" />
          </font>
