+++ src/main/Application/ForProductsController.java
@@ -1,16 +1,19 @@
 // author: Daniil-Y
 
-package main.Controller;
+package main.Application;
 
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.ResourceBundle;
+
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Label;
 import javafx.stage.Stage;
+import jdk.jfr.Event;
 import main.Data.Query;
 import javafx.scene.control.Button;
 
@@ -26,13 +29,12 @@ public class ControllerForProducts {
     private Label AllProducts;
     @FXML
     private Button cancelButton;
-    private String parentPath;
-    public void savePath(String path) {
-        this.parentPath = path;
+    @FXML
+    public void goToHome(ActionEvent e) {
+        Main.goToHome();
     }
     @FXML
     void initialize() {
-        assert AllProducts != null : "fx:id=\"AllProducts\" was not injected: check your FXML file 'forButtonProducts.fxml'.";
         try {
             ArrayList<ArrayList<String>> test = Query.getFullInformation("products");
             for(ArrayList<String> s: test){
@@ -41,11 +43,7 @@ public class ControllerForProducts {
         }catch(Exception e){
             e.printStackTrace();
         }
-        cancelButton.setOnAction(event -> {
-            Stage stage = (Stage) cancelButton.getScene().getWindow();
-            FXMLLoader loader = LoadXML.load(parentPath);
-            Parent root = loader.getRoot();
-            ((Stage) cancelButton.getScene().getWindow()).setScene(new Scene(root));
-        });
     }
+
+
 }
