+++ src/main/Application/ForReceipts.fxml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.text.Font?>
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Application.ControllerForReceipts">
+   <children>
+      <Label fx:id="ShowReceipts" layoutX="252.0" layoutY="102.0" />
+      <Button fx:id="cancelButton" onAction = "#goToHome" cancelButton="true" layoutX="28.0" layoutY="21.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="30.0" text="&lt;">
+         <font>
+            <Font name="Arial Black" size="15.0" />
+         </font>
+      </Button>
+   </children>
+</AnchorPane>
