+++ src/main/Application/sample.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.paint.LinearGradient?>
 <?import javafx.scene.paint.Stop?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Controller.Controller">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.Application.Controller">
    <children>
       <Button fx:id="show_all_products" layoutX="366.0" layoutY="92.0" mnemonicParsing="false" text="Products" textFill="#43a69e" textOverrun="WORD_ELLIPSIS" />
       <Button fx:id="show_all_receipts" layoutX="171.0" layoutY="92.0" mnemonicParsing="false" text="Receipts">
